+++ packages/react/src/__tests__/index-test.js
@@ -141,6 +141,7 @@ describe('Carbon Components React', () => {
         "RadioButtonSkeleton",
         "RadioTile",
         "Row",
+        "SIZES",
         "Search",
         "SearchSkeleton",
         "SecondaryButton",
@@ -182,6 +183,7 @@ describe('Carbon Components React', () => {
         "Switcher",
         "SwitcherDivider",
         "SwitcherItem",
+        "TYPES",
         "Tab",
         "TabContent",
         "TabList",
